+++ TODO/app/Http/Controllers/AuthController.php
@@ -8,46 +8,56 @@ use App\Http\Controllers\Controller;
 use App\User;
 use JWTAuth;
 use Validator;
+
 class AuthController extends Controller
 {
-    public function __construct()
-    {
-        $this->middleware('jwt.auth', ['except' => ['login']]);
-    }
-
-    /**
-     * Get a JWT via given credentials.
-     *
-     * @return \Illuminate\Http\JsonResponse
-     */
+    
     public function login(Request $request)
     {
         $validator = Validator::make($request->all(), [
             'email' => 'required|string|max:255|exists:users',
             'password'=> 'required'
         ]);
+
         $validator->validate();
-        
+
         $credentials = $request->only('email', 'password');
-        \Log::debug($credentials);
+
         try {
-            if (! $token = JWTAuth::attempt($credentials)) {
+            if (! $token = JWTAuth::attempt($credentials)) 
+            {
                 return response()->json(['error' => 'invalid_credentials'], 401);
             }
         } catch (JWTException $e) {
             return response()->json(['error' => 'could_not_create_token'], 500);
         }
+
         return response()->json(compact('token'));
-        //$user = User::findOrFail(1);
-        //\Log::debug($user);
-       // $token = JWTAuth::fromUser($user);
-        //return response()->json($token);
     }
 
     public function logout(Request $request)
     {
-       
         return response()->json(auth()->logout());
     }
 
+    public function register(Request $request)
+    {
+        $validator = Validator::make($request->all(), [
+            'email' => 'required|string|max:255|unique:users',
+            'name' => 'required',
+            'password'=> 'required'
+        ]);
+
+        $validator->validate();
+        
+        $user = User::create([
+            'name' => $request->get('name'),
+            'email' => $request->get('email'),
+            'password' => bcrypt($request->get('password')),
+        ]);
+        
+        $token = JWTAuth::fromUser($user);
+        
+        return response()->json(compact('token'));
+    }
 }
