+++ TODO/app/User.php
@@ -6,20 +6,14 @@ use Tymon\JWTAuth\Contracts\JWTSubject;
 use Illuminate\Notifications\Notifiable;
 use Illuminate\Foundation\Auth\User as Authenticatable;
 use Laravel\Passport\HasApiTokens;
+
 class User extends Authenticatable implements JWTSubject
 {
     use HasApiTokens, Notifiable;
-
-    /**
-     * The attributes that are mass assignable.
-     *
-     * @var array
-     */
-    protected $fillable = [
-        'id','name', 'email', 'password',
-    ];
-
     
+    protected $fillable = [
+        'name', 'email', 'password',
+    ];
 
     public function getJWTIdentifier()
     {
