+++ TODO/routes/web.php
@@ -10,6 +10,3 @@
 | contains the "web" middleware group. Now create something great!
 |
 */
-use App\Task;
-
-Route::post('/login','LoginController@login');
